refactor(工作台): 重构工作台首页项目概览和我创建的卡片调整人员概览交互调整
